Move RecordTask and CameraFramesDataStream from scan base to HarwdareSource and camera_base.
